Module merino_adm::remote_settings

AdM integration that uses the remote-settings provided data.

Modules

reqwest_client

An HTTP client implementation to use with the remote-settings-client.

soft_verifier

A custom signature verifier implementation that marks the verification as successful even in case of problems, but logs a warning about the failure.

Structs

AdmSuggestion

A suggestion record from AdM

AttachmentMeta

The metadata of an attachment that might be associated with a Remote Settings record.

NON_SPONSORED_IAB_CATEGORIES

RemoteSettingsSuggester

Make suggestions based on data in Remote Settings