Module merino_adm::remote_settings

AdM integration that uses the remote-settings provided data.

Structs

AdmSuggestion
A suggestion record from AdM

AttachmentMeta
The metadata of an attachment that might be associated with a Remote Settings record.

RemoteSettingsAttachmentsCapability
Remote Settings attachments capability

RemoteSettingsCapabilities
Remote Settings server capabilities

RemoteSettingsServerInfo
Remote Settings server info

RemoteSettingsSuggester
Make suggestions based on data in Remote Settings

SuggestRecord
A record stored in the Remote Settings quicksuggest collection.

Statics

REMOTE_SETTINGS_SERVER_INFO
A lazy version of the server settings for the default Remote Settings server. Should be initialized with RemoteSettingsServerInfo::fetch.